use integer linspace calls in examples

+++ examples/cpu/det.cpp
@@ -17,7 +17,7 @@ int main()
   len_t n = 2;
   
   cpumat<float> x(n, n);
-  x.fill_linspace(1.f, (float)n*n);
+  x.fill_linspace(1, n*n);
   
   x.info();
   x.print(0);
@@ -29,7 +29,7 @@ int main()
   
   n = 4;
   x.resize(n, n);
-  x.fill_linspace(1.f, (float)n*n);
+  x.fill_linspace(1, n*n);
   
   x.info();
   x.print(0);
